Add format support to CatalogSink and update methods
- Introduced a new format parameter in CatalogSink to specify the data
  format when writing.
- Updated the writeDataset method to apply the format if specified.
- Enhanced the CatalogSink constructor to accept format as an argument.
- Modified DataSinkBuilder to support format in the tables method.

+++ library-api/src/main/java/au/csiro/pathling/library/io/sink/CatalogSink.java
@@ -25,7 +25,6 @@ import au.csiro.pathling.library.io.SaveMode;
 import io.delta.tables.DeltaTable;
 import jakarta.annotation.Nonnull;
 import java.util.Optional;
-import java.util.function.UnaryOperator;
 import org.apache.spark.sql.DataFrameWriter;
 import org.apache.spark.sql.Dataset;
 import org.apache.spark.sql.Row;
@@ -46,6 +45,9 @@ public class CatalogSink implements DataSink {
   @Nonnull
   private final Optional<String> schema;
 
+  @Nonnull
+  private final Optional<String> format;
+
   /**
    * Constructs a CatalogSink with the specified PathlingContext and default import mode.
    *
@@ -55,6 +57,7 @@ public class CatalogSink implements DataSink {
     this.context = context;
     this.saveMode = SaveMode.ERROR_IF_EXISTS; // Default import mode
     this.schema = Optional.empty(); // Schema not specified
+    this.format = Optional.empty(); // Format not specified
   }
 
   /**
@@ -67,6 +70,7 @@ public class CatalogSink implements DataSink {
     this.context = context;
     this.saveMode = saveMode;
     this.schema = Optional.empty(); // Schema not specified
+    this.format = Optional.empty(); // Format not specified
   }
 
   /**
@@ -81,6 +85,23 @@ public class CatalogSink implements DataSink {
     this.context = context;
     this.saveMode = saveMode;
     this.schema = Optional.of(schema);
+    this.format = Optional.empty(); // Format not specified
+  }
+
+  /**
+   * Constructs a CatalogSink with the specified PathlingContext, import mode, schema, and format.
+   *
+   * @param context the PathlingContext to use
+   * @param saveMode the SaveMode to use when writing data
+   * @param schema the schema to qualify the table names, if any
+   * @param format the format to use when writing data
+   */
+  public CatalogSink(@Nonnull final PathlingContext context, @Nonnull final SaveMode saveMode,
+      @Nonnull final String schema, @Nonnull final String format) {
+    this.context = context;
+    this.saveMode = saveMode;
+    this.schema = Optional.of(schema);
+    this.format = Optional.of(format);
   }
 
   @Override
@@ -91,18 +112,20 @@ public class CatalogSink implements DataSink {
 
       switch (saveMode) {
         case ERROR_IF_EXISTS ->
-            writeDataset(dataset, tableName, org.apache.spark.sql.SaveMode.ErrorIfExists,
-                UnaryOperator.identity());
+            writeDataset(dataset, tableName, org.apache.spark.sql.SaveMode.ErrorIfExists);
         case OVERWRITE -> {
-          // This is to work around a bug relating to Delta tables not being able to be overwritten,
-          // due to their inability to handle the truncate operation that Spark performs when
-          // overwriting a table.
-          context.getSpark().sql("DROP TABLE IF EXISTS " + tableName);
-          writeDataset(dataset, tableName, org.apache.spark.sql.SaveMode.ErrorIfExists,
-              w -> w.option("overwriteSchema", "true"));
+          if (format.isPresent() && "delta".equals(format.get())) {
+            // This is to work around a bug relating to Delta tables not being able to be overwritten,
+            // due to their inability to handle the truncate operation that Spark performs when
+            // overwriting a table.
+            context.getSpark().sql("DROP TABLE IF EXISTS " + tableName);
+            writeDataset(dataset, tableName, org.apache.spark.sql.SaveMode.ErrorIfExists);
+          } else {
+            // Use standard overwrite for non-Delta formats.
+            writeDataset(dataset, tableName, org.apache.spark.sql.SaveMode.Overwrite);
+          }
         }
-        case APPEND -> writeDataset(dataset, tableName, org.apache.spark.sql.SaveMode.Append,
-            UnaryOperator.identity());
+        case APPEND -> writeDataset(dataset, tableName, org.apache.spark.sql.SaveMode.Append);
         case MERGE -> {
           if (DeltaTable.isDeltaTable(tableName)) {
             // If the table already exists, merge the data in.
@@ -110,20 +133,21 @@ public class CatalogSink implements DataSink {
             merge(table, dataset);
           } else {
             // If the table does not exist, create it.
-            writeDataset(dataset, tableName, org.apache.spark.sql.SaveMode.ErrorIfExists,
-                UnaryOperator.identity());
+            writeDataset(dataset, tableName, org.apache.spark.sql.SaveMode.ErrorIfExists);
           }
         }
       }
     }
   }
 
-  private static void writeDataset(@Nonnull final Dataset<Row> dataset,
-      @Nonnull final String tableName, @Nonnull final org.apache.spark.sql.SaveMode saveMode,
-      @Nonnull final UnaryOperator<DataFrameWriter<Row>> writerModifier) {
+  private void writeDataset(@Nonnull final Dataset<Row> dataset,
+      @Nonnull final String tableName, @Nonnull final org.apache.spark.sql.SaveMode saveMode) {
     final DataFrameWriter<Row> writer = dataset.write()
         .mode(saveMode);
-    writerModifier.apply(writer);
+    
+    // Apply format if specified
+    format.ifPresent(writer::format);
+    
     writer.saveAsTable(tableName);
   }
